# Make slices parsing faster by pre-allocation