--- a/pkg/meta/redis_test.go
+++ b/pkg/meta/redis_test.go
@@ -25,7 +25,7 @@ import (
 	"github.com/juicedata/juicefs/pkg/utils"
 )
 
-func BenchmarkReadSlices(b *testing.B) {
+func encodeSlices(size int) []string {
 	w := utils.NewBuffer(24)
 	w.Put32(0)
 	w.Put64(1014)
@@ -33,17 +33,34 @@ func BenchmarkReadSlices(b *testing.B) {
 	w.Put32(0)
 	w.Put32(122)
 	v := string(w.Bytes())
-	var vals [128]string
+	vals := make([]string, size)
 	for i := range vals {
 		vals[i] = v
 	}
-	b.ResetTimer()
-	var slices []*slice
-	for i := 0; i < b.N; i++ {
-		slices = readSlices(vals[:])
+	return vals
+}
+
+func BenchmarkReadSlices(b *testing.B) {
+	cases := []struct {
+		desc string
+		size int
+	}{
+		{"small", 4},
+		{"mid", 64},
+		{"large", 1024},
 	}
-	if len(slices) != len(vals) {
-		b.Fail()
+	for _, c := range cases {
+		b.Run(c.desc, func(b *testing.B) {
+			vals := encodeSlices(c.size)
+			b.ResetTimer()
+			var slices []*slice
+			for i := 0; i < b.N; i++ {
+				slices = readSlices(vals)
+			}
+			if len(slices) != len(vals) {
+				b.Fail()
+			}
+		})
 	}
 }