## Before

```
BenchmarkReadSlices/small-16             3035296               400 ns/op             352 B/op          9 allocs/op
BenchmarkReadSlices/small-16             3063632               382 ns/op             352 B/op          9 allocs/op
BenchmarkReadSlices/small-16             3184216               383 ns/op             352 B/op          9 allocs/op
BenchmarkReadSlices/mid-16                227870              5467 ns/op            5632 B/op        129 allocs/op
BenchmarkReadSlices/mid-16                221240              5342 ns/op            5632 B/op        129 allocs/op
BenchmarkReadSlices/mid-16                220996              5185 ns/op            5632 B/op        129 allocs/op
BenchmarkReadSlices/large-16               14389             84043 ns/op           90112 B/op       2049 allocs/op
BenchmarkReadSlices/large-16               14252             83926 ns/op           90112 B/op       2049 allocs/op
BenchmarkReadSlices/large-16               13748             85014 ns/op           90112 B/op       2049 allocs/op
```

## After

```
BenchmarkReadSlices/small-16             3990091               301 ns/op             320 B/op          6 allocs/op
BenchmarkReadSlices/small-16             3838144               304 ns/op             320 B/op          6 allocs/op
BenchmarkReadSlices/small-16             3875310               325 ns/op             320 B/op          6 allocs/op
BenchmarkReadSlices/mid-16                301630              3684 ns/op            5248 B/op         66 allocs/op
BenchmarkReadSlices/mid-16                330322              3584 ns/op            5248 B/op         66 allocs/op
BenchmarkReadSlices/mid-16                326588              3623 ns/op            5248 B/op         66 allocs/op
BenchmarkReadSlices/large-16               21490             55500 ns/op           81920 B/op       1026 allocs/op
BenchmarkReadSlices/large-16               21813             56052 ns/op           81920 B/op       1026 allocs/op
BenchmarkReadSlices/large-16               21601             56190 ns/op           81920 B/op       1026 allocs/op
```